requirements.txt
EssentialCV
Pillow
scipy
sklearn